+++ public/index.html
@@ -33,6 +33,6 @@
   </section>
 </footer>
 
-<script src="../src/main.js"></script>
+<script src="./bundle.js"></script>
 </body>
 </html>
