+++ src/view/constants.js
@@ -1,0 +1,9 @@
+export const FilmCardCount = {
+  DEFAULT: 5,
+  EXTRA: 2
+};
+
+export const EmojiSize = {
+  SMALL: 30,
+  LARGE: 55,
+};
