+++ src/view/container.js
@@ -1,0 +1,3 @@
+export const createContainerTemplate = () => {
+  return `<section class="films"></section>`;
+};
