+++ src/view/emoji.js
@@ -1,0 +1,5 @@
+export const createEmojiTemplate = (emoji, size) => {
+  return (
+    `<img src="./images/emoji/${emoji}.png" width="${size}" height="${size}" alt="emoji-${emoji}">`
+  );
+};
