+++ src/view/film-details.js
@@ -1,0 +1,62 @@
+export const createFilmDetaisTemplate = () => {
+  return (
+    `<div class="film-details__info-wrap">
+      <div class="film-details__poster">
+        <img class="film-details__poster-img" src="./images/posters/the-great-flamarion.jpg" alt="">
+
+        <p class="film-details__age">18+</p>
+      </div>
+
+      <div class="film-details__info">
+        <div class="film-details__info-head">
+          <div class="film-details__title-wrap">
+            <h3 class="film-details__title">The Great Flamarion</h3>
+            <p class="film-details__title-original">Original: The Great Flamarion</p>
+          </div>
+
+          <div class="film-details__rating">
+            <p class="film-details__total-rating">8.9</p>
+          </div>
+        </div>
+
+        <table class="film-details__table">
+          <tr class="film-details__row">
+            <td class="film-details__term">Director</td>
+            <td class="film-details__cell">Anthony Mann</td>
+          </tr>
+          <tr class="film-details__row">
+            <td class="film-details__term">Writers</td>
+            <td class="film-details__cell">Anne Wigton, Heinz Herald, Richard Weil</td>
+          </tr>
+          <tr class="film-details__row">
+            <td class="film-details__term">Actors</td>
+            <td class="film-details__cell">Erich von Stroheim, Mary Beth Hughes, Dan Duryea</td>
+          </tr>
+          <tr class="film-details__row">
+            <td class="film-details__term">Release Date</td>
+            <td class="film-details__cell">30 March 1945</td>
+          </tr>
+          <tr class="film-details__row">
+            <td class="film-details__term">Runtime</td>
+            <td class="film-details__cell">1h 18m</td>
+          </tr>
+          <tr class="film-details__row">
+            <td class="film-details__term">Country</td>
+            <td class="film-details__cell">USA</td>
+          </tr>
+          <tr class="film-details__row">
+            <td class="film-details__term">Genres</td>
+            <td class="film-details__cell">
+              <span class="film-details__genre">Drama</span>
+              <span class="film-details__genre">Film-Noir</span>
+              <span class="film-details__genre">Mystery</span></td>
+          </tr>
+        </table>
+
+        <p class="film-details__film-description">
+          The film opens following a murder at a cabaret in Mexico City in 1936, and then presents the events leading up to it in flashback. The Great Flamarion (Erich von Stroheim) is an arrogant, friendless, and misogynous marksman who displays his trick gunshot act in the vaudeville circuit. His show features a beautiful assistant, Connie (Mary Beth Hughes) and her drunken husband Al (Dan Duryea), Flamarion's other assistant. Flamarion falls in love with Connie, the movie's femme fatale, and is soon manipulated by her into killing her no good husband during one of their acts.
+        </p>
+      </div>
+    </div>`
+  );
+};
