+++ src/view/footer-statistics.js
@@ -1,0 +1,3 @@
+export const createFooterStatisticsTemplate = () => {
+  return `<p>130 291 movies inside</p>`;
+};
